requirements.txt
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.64.0
segmentation_models_pytorch>=0.2.1
opencv-python>=4.6.0
albumentations>=1.0.3
pyyaml